feat:Update review

See merge request s08-webmobile1-sub2/S08P12D105!125

+++ Back-End/src/main/java/com/ssafy/style/data/dto/ReviewDto.java
@@ -20,7 +20,7 @@ public class ReviewDto {
     @ApiModelProperty(value = "컨설턴트아이디")
     private String consultantId;
     @ApiModelProperty(value = "리뷰 점수")
-    private Integer reviewScore;
+    private Double reviewScore;
     @ApiModelProperty(value = "리뷰 내용")
     private String reviewContent;
     @ApiModelProperty(value = "작성 날짜")
